+++ COMP371_A3/Sphere.h
@@ -6,14 +6,14 @@ class Sphere :
 	public Primitive
 {
 	glm::vec3 m_centre;
-	double m_radius;
+	float m_radius;
 
 public:
-	Sphere(const glm::vec3 m_centre, const double radius, glm::vec3 ambientColour, const glm::vec3 diffuseColour, const glm::vec3 specularColour, const float shininess);
+	Sphere(const glm::vec3 m_centre, const float radius, glm::vec3 ambientColour, const glm::vec3 diffuseColour, const glm::vec3 specularColour, const float shininess);
 	glm::vec3 centre() const;
-	double radius() const;
+	float radius() const;
 	void setCentre(const glm::vec3 centre);
-	void setRadius(const double radius);
-	virtual std::pair<bool, double> intersection(const Ray ray);
+	void setRadius(const float radius);
+	virtual std::pair<bool, float> intersection(const Ray ray);
 };
