+++ COMP371_A3/primitive.h
@@ -33,5 +33,5 @@ public:
 	void setDiffuseColour(const glm::vec3 diffuseColour);
 	void setSpecularColour(const glm::vec3 specularColour);
 	void setShininess(const float shininess);
-	virtual std::pair<bool, double> intersection(const Ray ray) = 0;
+	virtual std::pair<bool, float> intersection(const Ray ray) = 0;
 };
